requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
crispy-bootstrap5==0.4
Cython==0.29.23
Django==3.2.4
django-crispy-forms==1.12.0
django-fastdev==1.0.2
Flask==2.0.1
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonpickle==2.0.0
MarkupSafe==2.0.1
packaging==20.9
pluggy==0.13.1
protobuf==3.11.3
psycopg2-binary==2.9.1
py==1.10.0
pycrunch-trace==0.1.6
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.1
whitenoise==5.3.0
